redirect method with acl lookup
