+++ apps/xyz/mod/user/cookie.js
@@ -12,7 +12,7 @@ Exports the [user] cookie method for the /api/user/cookie route.
 */
 
 import jwt from 'jsonwebtoken';
-import { getRedirect, setRedirect } from '../utils/redirect.js';
+import { setRedirect } from '../utils/redirect.js';
 import acl from './acl.js';
 import login from './login.js';
